Make _send_request "public"
This is cosmetic, but it changes _send_request to send_request which
has the implication that it's "part of the API" now.

The use case for this is that I'm creating a CurlyElasticSearch class
that wraps the send_request method and logs the ES requests in a
"curl-ified" way making it really easy to debug problems by copy
and pasting the curl lines.